+++ src/settings/SettingsTab.ts
@@ -32,6 +32,8 @@ import makeOctaviaSettings from "./per-lang/makeOctaveSettings";
 import makeMaximaSettings from "./per-lang/makeMaximaSettings";
 import makeApplescriptSettings from "./per-lang/makeApplescriptSettings";
 import makeZigSettings from "./per-lang/makeZigSettings";
+import makeOCamlSettings from "./per-lang/makeOCamlSettings";
+import makeSwiftSettings from "./per-lang/makeSwiftSettings";
 
 
 /**
@@ -198,6 +200,9 @@ export class SettingsTab extends PluginSettingTab {
 		// ========== Scala ===========
 		makeScalaSettings(this, this.makeContainerFor("scala"));
 
+		// ========== Swift ===========
+		makeSwiftSettings(this, this.makeContainerFor("swift"));
+
 		// ========== Racket ===========
 		makeRacketSettings(this, this.makeContainerFor("racket"));
 
@@ -221,6 +226,9 @@ export class SettingsTab extends PluginSettingTab {
 
 		// ========== Zig ============
 		makeZigSettings(this, this.makeContainerFor("zig"));
+
+		// ========== OCaml ============
+		makeOCamlSettings(this, this.makeContainerFor("ocaml"));
 
 		this.focusContainer(this.plugin.settings.lastOpenLanguageTab || canonicalLanguages[0]);
 	}