+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 
+## [1.11.0]
+### Added
+- Support for OCaml (Thanks to @nieomylnieja)
+- Support for Swift (Thanks to @ihomway)
+
+### Changed
+- Improved support for C compiler gcc (Thanks to @melo-afk)
+
+
 ## [1.10.0]
 ### Added
 - Support for zig (Thanks to @slar)